requirements.txt
numpy>=1.21.0
Pillow>=9.0.0 